TODO rename g_sched_resp to something else Call into scheduler directly when receiving slot indication
The previous designs seems to do:
loop {
  poll_ind_queue()
  if (msg)
    scheduler;

  pselect() on messages;
  handle_message {
    if slot_ind
      put_ind_queue()
  }
}

So basically, we artificially put a queue for slot indications in the
middle, but still handle it in the same thread(!). This for some reason
results in a big slow down if the PNF runs faster.

Now, we just do pselect(), waiting for messages. We handle the slot
indication immediately (the schedule should take some microseconds),
then return to pselect().