Remove reestablishment on target cell
Reestablishment on target cell during handover cannot be handled by the
gNB (yet). It is the same as with a normal reestablishment (not during
handover) on a different cell: that cell has no notion of the old UE, so
we cannot simply indicate the old UE to reestablish.

Instead, we would need to trigger a UE context setup message, including
the old cellGroupConfig, so that the new cell can setup reception for
this UE.