Handle NGAP Initial Context Setup Request if AS security is activated
To me, TS 38.413 is not really clear whether the gNB should be prepared
to handle multiple Initial Context Setup Requests or not [1]. Before
this commit, the gNB would simply assert:

    Assertion (!ue_p->as_security_active) failed!
    In rrc_gNB_generate_SecurityModeCommand() /path/to/openairinterface5g/openair2/RRC/NR/rrc_gNB.c:2590
    logic error: security already active

To more gracefully handle this, check if AS security is activated. If
it is not, we do the same as before: trigger Security Mode Command,
potentially store PDU sessions to setup, and also the NAS PDU that we
might need to send.

If AS security is already activated, two things can happen:

1. If the keys change: I don't know how to handle this, and 38.331 reads
   as if SecurityModeCommand should only happen at the very beginning,
   and we might not retrigger SecurityModeCommand because DRBs might be
   setup. So, reject the message.

2. If the keys don't change: AS security is up, and we should also have
   UE capabilities. We can then either set up the PDU sessions, or if
   there are none, directly acknowledge the Initial context setup.

[1] 38.413 section 8.3.1.1:
> The AMF may initiate the Initial Context Setup procedure if a
> UE-associated logical NG-connection exists

This means it can always happen?